Update inline docs for fields.Raw
Updates the documentation for `fields.Raw` to reflect that overriding `format` is only appropriate when object keys exist, and that `output` should be overridden when data is being created that does not depend on the output key. 

The wording of this could perhaps be improved, but as a new flask-restful user, I struggled with this because the existing documentation (both here and in docs/fields.rst) was misleading.