Fixes handling of NotAcceptable exceptions
When default_mimetype is set to None, it is now overridden
by the first representation that we support, such that the
NotAcceptable (406) error content itself can be returned
in that format.

This fixes a bug wherein that only occurred on the first
request, because it was setting default_mimetype on the
whole API rather than just overriding it for one request.